weatherContext added & spinner moved to its own component
